requirements.txt
Pillow>=1.1.7
PyQt5>=5.12
PyQtWebEngine>=5.12
